+++ openquake/hazardlib/gsim/base.py
@@ -26,12 +26,11 @@ import inspect
 import warnings
 import functools
 import numpy
-from scipy.special import ndtr
-from scipy.stats import norm
 
 from openquake.baselib.general import DeprecationWarning
 from openquake.baselib.performance import compile
 from openquake.hazardlib import const
+from openquake.hazardlib.stats import _truncnorm_sf
 from openquake.hazardlib.gsim.coeffs_table import CoeffsTable
 from openquake.hazardlib.contexts import KNOWN_DISTANCES
 from openquake.hazardlib.contexts import *  # for backward compatibility
@@ -100,100 +99,6 @@ def _get_poes(mean_std, loglevels, truncation_level):
     delta = _get_delta(mean_std, loglevels.array, truncation_level,
                        loglevels.L1)
     return _truncnorm_sf(truncation_level, delta)
-
-
-def _get_poes_site(mean_std, loglevels, truncation_level, ampfun, ctxs):
-    """
-    NOTE: this works for a single site
-
-    :param mean_std:
-        See :function:`openquake.hazardlib.gsim.base.get_poes`
-    :param loglevels:
-        Intensity measure level per intensity measure type. See
-        :function:`openquake.hazardlib.gsim.base.get_poes`
-    :param truncation_level:
-        The level of truncation of the normal distribution of ground-motion
-        on rock
-    :param ampl:
-        Site amplification function instance of
-        :class:openquake.hazardlib.site_amplification.AmpFunction
-    :param ctxs:
-        Context objects with attributes .mag, .sites, .rrup
-    """
-    # Mean and std of ground motion for the IMTs considered in this analysis
-    # C - Number of contexts
-    # L - Number of intensity measure levels
-    mean, stddev = mean_std  # shape (C, M)
-    C, L = mean.shape[1], loglevels.size
-    for ctx in ctxs:
-        assert len(ctx.sids) == 1  # 1 site
-    M = len(loglevels)
-    L1 = L // M
-
-    # This is the array where we store the output results i.e. poes on soil
-    out_s = numpy.zeros((C, L))
-
-    # `nsamp` is the number of IMLs per IMT used to compute the hazard on rock
-    # while 'L' is total number of ground-motion values
-    nsamp = 40
-
-    # Compute the probability of exceedance for each in intensity
-    # measure type IMT
-    sigma = ampfun.get_max_sigma()
-    mags = [ctx.mag for ctx in ctxs]
-    rrups = [ctx.rrup for ctx in ctxs]
-    ampcode = ctxs[0].sites['ampcode'][0]
-    for m, imt in enumerate(loglevels):
-
-        # Get the values of ground-motion used to compute the probability
-        # of exceedance on soil.
-        soillevels = loglevels[imt]  # shape L1
-
-        # Here we set automatically the IMLs that will be used to compute
-        # the probability of occurrence of GM on rock within discrete
-        # intervals
-        ll = numpy.linspace(min(soillevels) - sigma * 4.,
-                            max(soillevels) + sigma * 4.,
-                            num=nsamp)
-
-        # Calculate for each ground motion interval the probability
-        # of occurrence on rock for all the sites
-        for iml_l, iml_u in zip(ll[:-1], ll[1:]):
-
-            # Set the arguments of the truncated normal distribution
-            # function
-            if truncation_level == 0:
-                out_l = iml_l <= mean[m]
-                out_u = iml_u <= mean[m]
-            else:
-                out_l = (iml_l - mean[m]) / stddev[m]
-                out_u = (iml_u - mean[m]) / stddev[m]
-
-            # Probability of occurrence on rock
-            pocc_rock = (_truncnorm_sf(truncation_level, out_l) -
-                         _truncnorm_sf(truncation_level, out_u))  # shape C
-
-            # Skipping cases where the pocc on rock is negligible
-            if numpy.all(pocc_rock < 1e-10):
-                continue
-
-            # Ground-motion value in the middle of each interval
-            iml_mid = (numpy.exp(iml_l) + numpy.exp(iml_u)) / 2.
-
-            # Get mean and std of the amplification function for this
-            # magnitude, distance and IML
-            median_af, std_af = ampfun.get_mean_std(  # shape C
-                ampcode, imt, iml_mid, mags, rrups)
-
-            # Computing the probability of exceedance of the levels of
-            # ground-motion loglevels on soil
-            logaf = numpy.log(numpy.exp(soillevels) / iml_mid)  # shape L1
-            for li in range(L1):
-                poex_af = 1. - norm.cdf(
-                    logaf[li], numpy.log(median_af), std_af)  # shape C
-                out_s[:, m * L1 + li] += poex_af * pocc_rock  # shape C
-
-    return out_s
 
 
 OK_METHODS = 'compute get_mean_and_stddevs get_poes set_parameters'
@@ -456,62 +361,6 @@ class GroundShakingIntensityModel(metaclass=MetaGSIM):
         if self._toml:
             return self._toml
         return '[%s]' % self.__class__.__name__
-
-
-def _truncnorm_sf(truncation_level, values):
-    """
-    Survival function for truncated normal distribution.
-
-    Assumes zero mean, standard deviation equal to one and symmetric
-    truncation.
-
-    :param truncation_level:
-        Positive float number representing the truncation on both sides
-        around the mean, in units of sigma, or None, for non-truncation
-    :param values:
-        Numpy array of values as input to a survival function for the given
-        distribution.
-    :returns:
-        Numpy array of survival function results in a range between 0 and 1.
-
-    >>> from scipy.stats import truncnorm
-    >>> truncnorm(-3, 3).sf(0.12345) == _truncnorm_sf(3, 0.12345)
-    True
-    >>> from scipy.stats import norm
-    >>> norm.sf(0.12345) == _truncnorm_sf(None, 0.12345)
-    True
-    """
-    if truncation_level == 0:
-        return values
-
-    if truncation_level is None:
-        return ndtr(- values)
-
-    # notation from http://en.wikipedia.org/wiki/Truncated_normal_distribution.
-    # given that mu = 0 and sigma = 1, we have alpha = a and beta = b.
-
-    # "CDF" in comments refers to cumulative distribution function
-    # of non-truncated distribution with that mu and sigma values.
-
-    # assume symmetric truncation, that is ``a = - truncation_level``
-    # and ``b = + truncation_level``.
-
-    # calculate CDF of b
-    phi_b = ndtr(truncation_level)
-
-    # calculate Z as ``Z = CDF(b) - CDF(a)``, here we assume that
-    # ``CDF(a) == CDF(- truncation_level) == 1 - CDF(b)``
-    z = phi_b * 2 - 1
-
-    # calculate the result of survival function of ``values``,
-    # and restrict it to the interval where probability is defined --
-    # 0..1. here we use some transformations of the original formula
-    # that is ``SF(x) = 1 - (CDF(x) - CDF(a)) / Z`` in order to minimize
-    # number of arithmetic operations and function calls:
-    # ``SF(x) = (Z - CDF(x) + CDF(a)) / Z``,
-    # ``SF(x) = (CDF(b) - CDF(a) - CDF(x) + CDF(a)) / Z``,
-    # ``SF(x) = (CDF(b) - CDF(x)) / Z``.
-    return ((phi_b - ndtr(values)) / z).clip(0.0, 1.0)
 
 
 def to_distribution_values(vals, imt):
